+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### [0.8.1](Unreleased)
 
 #### Fixed
+- fix: add o2o field does not create constraint when migrating. (#396)
 - Migration with duplicate renaming of columns in some cases. (#395)
 - fix: intermediate table for m2m relation not created. (#394)
 - Migrate add m2m field with custom through generate duplicated table. (#393)
@@ -16,6 +17,7 @@
 - Fix configuration file reading error when containing Chinese characters. (#286)
 - sqlite: failed to create/drop index. (#302)
 - PostgreSQL: Cannot drop constraint after deleting or rename FK on a model. (#378)
+- Fix create/drop indexes in every migration. (#377)
 - Sort m2m fields before comparing them with diff. (#271)
 
 #### Changed